fixes #4066 (NewStory: Новая проблема со скроллингом)

+++ plugins/NewStory/src/history_control.cpp
@@ -521,7 +521,7 @@ void NewstoryListData::HitTotal(int yCurr, int yTotal)
 	while (i < totalCount && y > 0) {
 		auto *pItem = GetItem(i++);
 		if (!pItem->m_bLoaded) {
-			i = (totalCount * yCurr) / yTotal;
+			i = totalCount * (double(yCurr) / double(yTotal));
 			y = 0;
 			break;
 		}
